기능: 스킬 속성 관련 최대최소값 보정 기능 추가
- 수정 관리 이용 쉽도록 CommonDefine에 최대치 설정

+++ Assets/Script/Game/Data/ObjectData/ProjectileObjectData.cs
@@ -9,7 +9,7 @@ public class ProjectileObjectData : BaseObject, IPoolable
     public IPool Pool { get; set; }
 
     public float speed;
-    public int damage;
+    public float damage;
     public LayerMask targetLayer;
     public ISkillBehavior behavior; // 탄환 추가 행동
